tests/wpt/web-platform-tests/css/css-contain/content-visibility/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Content Visibility: content-visibility inheritance</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<meta name="assert" content="content-visibility does not inherit">
<meta name="assert" content="content-visibility has initial value 'visible'">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<script>
assert_not_inherited("content-visibility", "visible", "auto");
</script>
</body>
</html>
